requirements.txt
langchain
openai>=0.27.0
chromadb>=0.3.23
psycopg2-binary
sqlalchemy
pandas
uvicorn
fastapi
python-dotenv
tqdm